fix: adjust styles
