+++ src/modules/favoriteMovies/FavoriteMovies.page.jsx
@@ -1,22 +1,15 @@
 import { useEffect } from "react";
 
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 
 import ErrorState from "../../shared/components/errorState/ErrorState.jsx";
 import LoadingState from "../../shared/components/loadingState/LoadingState.jsx";
 import MovieGrid from "../../shared/components/movieGrid/MovieGrid.jsx";
 
-import { getFavoriteMovies } from "./favoriteMoviesSlice.js";
-
 const FavoriteMoviesPage = () => {
-  const dispatch = useDispatch();
   const { favoriteMovies, status, error } = useSelector(
     (state) => state.favoriteMovies,
   );
-
-  useEffect(() => {
-    dispatch(getFavoriteMovies());
-  }, [dispatch]);
 
   if (status === "loading") {
     return <LoadingState />;
