+++ src/modules/watchLaterMovies/WatchLaterMovies.page.jsx
@@ -1,22 +1,13 @@
-import { useEffect } from "react";
-
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 
 import ErrorState from "../../shared/components/errorState/ErrorState.jsx";
 import LoadingState from "../../shared/components/loadingState/LoadingState.jsx";
 import MovieGrid from "../../shared/components/movieGrid/MovieGrid.jsx";
 
-import { getWatchLaterMovies } from "./watchLaterMoviesSlice.js";
-
 const WatchLaterPage = () => {
-  const dispatch = useDispatch();
   const { watchLaterMovies, status, error } = useSelector(
     (state) => state.watchLaterMovies,
   );
-
-  useEffect(() => {
-    dispatch(getWatchLaterMovies());
-  }, [dispatch]);
 
   if (status === "loading") {
     return <LoadingState />;
