+++ src/shared/components/emptyState/EmptyState.scss
@@ -12,7 +12,7 @@
     background-color: white;
     color: var(--secondary-color);
     &:hover {
-      opacity: 80%;
+      opacity: 60%;
     }
   }
 }
