+++ src/shared/components/movieCard/MovieCard.scss
@@ -79,7 +79,7 @@
       color: white;
       padding-block: 0.5rem;
       &:hover {
-        background-color: rgba(255, 255, 255, 0.1);
+        opacity: 60%;
       }
     }
   }
