+++ src/shared/utils/renderWithProviders.js
@@ -1,33 +1,28 @@
 import React from "react";
-
 import { configureStore } from "@reduxjs/toolkit";
-import { setupListeners } from "@reduxjs/toolkit/dist/query";
 import { render } from "@testing-library/react";
 import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 
-import moviesSlice from "../../data/popularMoviesSlice";
-import favoriteMoviesSlice from "../../modules/favoriteMovies/favoriteMoviesSlice";
-import watchLaterMoviesSlice from "../../modules/watchLaterMovies/watchLaterMoviesSlice";
+import popularMoviesSliceReducer from "../../modules/popularMovies/popularMoviesSlice";
+import favoriteMoviesSliceReducer from "../../modules/favoriteMovies/favoriteMoviesSlice";
+import watchLaterMoviesSliceReducer from "../../modules/watchLaterMovies/watchLaterMoviesSlice";
 
-// @Todo: update to configureTestStore @claude
+const configureTestStore = (preloadedState = {}) => {
+  return configureStore({
+    reducer: {
+      popularMovies: popularMoviesSliceReducer,
+      favorite: favoriteMoviesSliceReducer,
+      watchLater: watchLaterMoviesSliceReducer,
+    },
+    preloadedState,
+  });
+};
+
 export function renderWithProviders(
   ui,
-  {
-    preloadedState = {},
-    store = configureStore({
-      reducer: {
-        popularMovies: moviesSlice.reducer,
-        favorite: favoriteMoviesSlice.reducer,
-        watchLater: watchLaterMoviesSlice.reducer,
-      },
-      preloadedState,
-    }),
-    ...renderOptions
-  } = {},
+  { preloadedState = {}, store = configureTestStore(preloadedState), ...renderOptions } = {}
 ) {
-  setupListeners(store.dispatch);
-
   function Wrapper({ children }) {
     return (
       <Provider store={store}>
@@ -38,3 +33,5 @@ export function renderWithProviders(
 
   return { store, ...render(ui, { wrapper: Wrapper, ...renderOptions }) };
 }
+
+export { configureTestStore };
